Check if exchange rate exists on all endpoints

+++ src/main/java/com/uniara/rest/service/ExchangeRateService.java
@@ -2,6 +2,7 @@ package com.uniara.rest.service;
 
 import com.uniara.rest.domain.ExchangeRate;
 import com.uniara.rest.exceptions.AlreadyExistsException;
+import com.uniara.rest.exceptions.NotFoundException;
 import com.uniara.rest.repository.ExchangeRateRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,8 +19,8 @@ public class ExchangeRateService {
     private ExchangeRateRepository exchangeRateRepository;
 
     public void create(final ExchangeRate exchangeRate) throws AlreadyExistsException {
-        if(findBySymbol(exchangeRate.getSymbol()) != null) {
-            throw new AlreadyExistsException();
+        if(this.exchangeRateRepository.exists(exchangeRate.getSymbol())) {
+           throw new AlreadyExistsException();
         }
         exchangeRateRepository.save(exchangeRate);
     }
@@ -28,11 +29,17 @@ public class ExchangeRateService {
         return exchangeRateRepository.findAll();
     }
 
-    public ExchangeRate findBySymbol(final String symbol) {
+    public ExchangeRate findBySymbol(final String symbol) throws NotFoundException {
+        if(!this.exchangeRateRepository.exists(symbol)) {
+            throw new NotFoundException();
+        }
         return this.exchangeRateRepository.findBySymbol(symbol);
     }
 
-    public boolean deleteBySymbol(final String symbol) {
+    public boolean deleteBySymbol(final String symbol) throws NotFoundException {
+        if(!this.exchangeRateRepository.exists(symbol)) {
+            throw new NotFoundException();
+        }
         return this.exchangeRateRepository.deleteBySymbol(symbol);
     }
 }
